move instruct checking
